fix(ci): Run the bindings check when build commands are updated (#2604)
# Motivation
The version of the signer was updated in #2592 but CI did not check the
bindings.

# Changes
- Include the build scripts in the list of files that cause the bindings
check to run. The build scripts include the external canister version
number.

Alternative considered:
- We could add the canister release number in `dfx.json`. This would be
outside the schema, but dfx has never objected to additional useful data
being stored in `dfx.json`. Theoretically the release number and the
location of the candid file are all that are needed to specify the
candid exactly and if those are both in `dfx.json` we can keep the list
of triggers small, leading to faster CI.
- I leave it up to the reviewer which they prefer. In the PR I have gone
for "conservative but slow" but would be very happy to use this instead.

# Tests

<!-- Please provide any information or screenshots about the tests that
have been done -->