refactor(frontend): Move source address in Send modal (#2633)
# Motivation

We want to reuse the `SendModal` component for Bitcoin transactions.

There is one part that is specific to IC, which is the address shown in
the form and review steps.

In this PR, I move this part up the chain so that we can pass later the
bitcoin address instead of the IC one.

This will also help me find pending transactions, which will be keyed by
address in a store and disable the form if any is present. This way, the
idea of pending transactions doesn't need to be specific to BTC.

# Changes

* Add prop `source` to `IcSendReview` and `IcSendForm` and use it
instead of `icrcAccountIdentifierText` store.
* Add prop `IcSendTokenWizard` and `SendWizard` to pass it down.
* Create local variable `source` using `icrcAccountIdentifierText` store
in `SendModal` and `IcSendModal` to then pass it down.

# Tests

Still working locally.