feat(frontend): Derived store to find pending transactions (#2652)
# Motivation

We want to display a warning messange and disable the send form if the
modal is for bitcoin and the user has pending transactions.

The bitcoin addresses are unique. Therefore, I wanted to take advantage
of that instead of checking selected networks or tokens.

The idea is to create a derived store creator that will return `true`,
`false` or `"loading"` depending on the address and whether the bitcoin
addresses have been loaded:

* Return `true` when the address is a bitcoin address and the pending
transactions array is not empty.
* Return `"loading"` if the addresses or the pending transactions are
not loaded. This is trickier because we also need to check the
`testnets`. If testnets are not enabled, then the addresses for testnet
and regtest won't be loaded.
* Return `false` if the pending transactions are present and empty.

This derived store will be used in `IcSendForm`:

```
let hasPendingTransactionsStore: Readable<boolean | "loading">;
$: hasPendingTransactionsStore = initHasPendingTransactionsStore(source);

//...

{#if $hasPendingTransactionsStore}
  <WarningMessage>{...}</>
```

# Changes

* New derived store creator `initHasPendingTransactions`.

# Tests

* Add tests for the new derived store.