feat(frontend): init signer allowance on sign-in before getting addresses (#2650)
# Motivation

We aim to introduce a new API that utilizes the `Signer` to load BTC and
ETH addresses. This requires the frontend to request, at most once per
session, that the signer provide an allowance for the client to derive
the addresses.

This pattern will need to be implemented for two use cases in Oisy, as
we differentiate between a fresh sign-in (with a blank client state) and
returning users who have cached addresses on their side (in IndexedDB).

This PR addresses the first use case. When the user signs in and the
'Loading' modal is displayed, we should initiate the signer allowance
before fetching the mainnet addresses, as we already do.

# Schema


![image](https://github.com/user-attachments/assets/4bfe5565-023e-4f4e-a644-7b93cd78bd23)

# Notes

This PR provides both services and usage examples for a comprehensive
review, given that the usage itself does not require many lines of code.

# Changes

- Create function `initSignerAllowance` which initialize the allowance
and sign-out the user in case of any error
- Consumes the new function in the `Loader` component

# Screenshot

<img width="1536" alt="Capture d’écran 2024-10-04 à 09 55 28"
src="https://github.com/user-attachments/assets/bc69ca7b-2fc1-48c4-bfd3-b632fdb8145b">