feat(frontend): init signer allowance before validating addresses (#2632)
# Motivation

We aim to introduce a new API that utilizes the `Signer` to load BTC and
ETH addresses. This requires the frontend to request, at most once per
session, that the signer provide an allowance for the client to derive
the addresses.

This pattern will need to be implemented for two use cases in Oisy, as
we differentiate between a fresh sign-in (with a blank client state) and
returning users who have cached addresses on their side (in IndexedDB).

The first use case has been implemented in PR #2650.

This PR addresses the second use case. When the user signs in to Oisy
with existing addresses saved in their browsers' chache. Adresses which
need to be validate with update calls in the background to ensure those
are still valid and matching users' principals.

# Schema


![image](https://github.com/user-attachments/assets/e9c322ad-45c8-4428-8860-1361748a74ba)

# Changes

- New global event `oisyValidateAddresses`. We do not need to save the
fact that the addresses need to be validated, therefore we do not need a
store. Plus, using an event allow some flexibility and is relatively
readable.
- Emit `oisyValidateAddresses` from the `Loader` when user is signed-in
without "splash modal" - i.e. when IDB addresses are used.
- Integrate `initSignerAllowance` and window listener for the event in
to `AddressGuard`.
- Update `AddressGuard` to prevent any race condition, i.e. validate
addresses when stores are updated and when the event is emitted. Note
that the services that validate the addresses takes care of performing a
unique update call when required - i.e. the related services can be
called multiple times.