feat(frontend): Add support for error in btc pending sent txs (#2704)
# Motivation

We want to show a different message to the user if loading the pending
sent transactions of bitcoin failed.

In this PR, I add the `null` type of the bitcoin pending sent
transactions store.

# Changes

* Add `null` in `BtcPendingSentTransactionsStoreData`.
* Add specific method to set it to `null`:
`setPendingTransactionsError`. This should allow us to easily change the
inner type.
* Support the new type in `initHasPendingSentTransactions` and return
`"error"`. This will also allow us to easily change the implementation
details of the error.

# Tests

* Test the new type in `btcPendingSentTransactionsStore` tests.
* Test the new type in `initHasPendingSentTransactions` tests.