fix(frontend): util formatToken was removing trailing zeros (#2717)
# Motivation

The util `formatToken` was not formatting correctly when the
`trailingZeros` prop was `true`. The behavior was the same as if it was
`false`. Meaning that all the numbers were formatted without trailing
zeros.

We adjust the util (we will need that prop in another PR) and we create
tests.

Note that we defaults `trailingZeros` to `false`, so that it keeps the
same behaviour as before. See image below, no changes:

<img width="1266" alt="Screenshot 2024-10-07 at 18 24 53"
src="https://github.com/user-attachments/assets/2171c252-f2ff-4bba-98c1-409479ff697c">