test(frontend): cleanup imports and mocks (#2722)
# Motivation

Both our mocks and imports of our tests would beneficit of a bit of
cleaning.

# Notes

This can probably be multiple PRs but, it all has the same goal:
cleaning up the tests. Therefore I allowed myself to provide a single
PR. Let me know if you rather like me to split it.

# Changes

- Replace relative imports with `$tests/...` 
- Setup support for `$tests`, which was already configured in `vitest`,
in `tsconfig.spec.json` - i.e. finish the configuration
- The Btc mocks weren^tprefixed with `mock...` while other mocks were
prefixed. So prefix the mocks.
- Since we can reuse the `mockBtcAdress` we can also extract the
`mockEthAdress` in a dedicated module.