style(frontend): change network switcher menu style [GIX-3086] (#2720)
# Motivation

We change the style of the dropdown for the Network selection. To do
that, I create a new `Dropdown` component to be re-used.

It is basically a refactoring of the previous `NetworksSwitcher`, but
breaking it down into smaller parts, to make it easier to change the
style eventually. We use Tailwind for the entire PR (note that there is
a repeated `blue-ribbon` color because the aim is to remove the RGB one
soon enough).

# Changes

- Create `DropdownButton` element with its specific style for each
state.
- Create `Dropdown` with `DropdownButton` and `items` slot.
- Wrap list of networks with `DropDown`.
- Remove Popover backdrop.
- Remove deprecated component.

# Tests


https://github.com/user-attachments/assets/33b50824-8168-4efd-ad00-3188086e0c0c



https://github.com/user-attachments/assets/cb027cd9-01ef-440c-8c86-544ce7ac6851