style(frontend): new ButtonAsLink for the footer [GIX-3049] (#2731)
# Motivation

We want to change the style of the icons in the footer. That is actually
a version of the `tertiary` button. So we adapt the code to use it.

# Changes

- Create a type for the button styles and use it in the button
components.
- Change the style of `tertiary` button class to have a generic style
and only specific after link.
- Create `ButtonAsLink` component that is basically the same as `Button`
but with <a>.
- Use `ButtonAsLink` in the `Footer`.
- Unrelated: we adjust the footer spacing.

# Tests

Only changed the footer. The rest is the same.

<img width="864" alt="Screenshot 2024-10-08 at 13 31 19"
src="https://github.com/user-attachments/assets/459e910f-665f-41e2-8303-8799b8fd0880">