feat(frontend): Add warning if btc pending transactions (#2774)
# Motivation

We want to avoid users double spending their utxos. Therefore, we
disable sending Bitcoin if there is any pending sent transaction.

In this PR, we show a warning and disable the "Send" button in the
review step.

# Changes

* Use `initHasPendingSentTransactions` to get the state of the btc
pending sent transactions, disable the button accordingly and show
warning.
* Add a slot in `SendReview` that can be used to add feedback to the
user.
* Add two new messages in i18n.

# Tests

Tested locally by filling up the store with the different scenarios.