feat(frontend): Add allowance method to Send services (#2812)
# Context

A partner encountered an issue not specific to Oisy, but somehow can be
food for thought for improvement.

It may happen that after the approval transaction is done, the transfer
transaction may fail for whatever reasons. In this case, the next
approval transaction will be rejected, since there is already an
approved amount. So, the next transfer transaction will work only if the
amount is below the one already approved. Otherwise, the user is blocked
until it consumes all the approved amount.

# Example

1. The user wants to convert 10 USDT to ckUSDT:
    1. `approve` transaction for 10 USDT --> _successful_
2. `transfer_from` transaction for 10 USDT --> _failed_ (for whatever
reason)
2. The user wants to convert 20 USDT to ckUSDT:
1. `approve` transaction for 20 USDT --> _failed_ (the previous approved
amount is still there)
2. `transfer_from` transaction for 20 USDT --> _failed_ (not enough
approved amount)
3. The user wants to convert 7 USDT to ckUSDT:
1. `approve` transaction for 7 USDT --> _failed_ (the previous approved
amount is still there)
2. `transfer_from` transaction for 7 USDT --> _successful_ (remaining
approved amount is 3 USDT)

# Practical example

We had a practical example with USDT. Please check the sender wallet
transaction history backwards from this one:
https://etherscan.io/tx/0x084432404a919c1ee720c906bb5c7a565b6e1fdcd5bb9da32049dddfa21ad23f

# Motivation

As first step to help the User with this: for the Send flow, it is
useful to know the allowance of the ERC20 smart contract for the Helper
smart contract. In this PR, we add the method in the provider and in the
services.

Consequently, in another PR, we will use it.