style(frontend): adjust main content to have a menu column [GIX-3050] (#2775)
# Motivation

We want to have a lateral column with the menu, but only for desktop and
tablets. So, this PR will first adjust the main content layout. We will
add the menu in another PR (there is a `TODO`).

# Changes

- Include Hero in the central part of the main content, and consequently
adjust its layout (spacing, transition, etc).
- Create a component for the MainLayout:
- On Desktop, it will show 3 columns (the menu on the left, the central
main content, and an empty column on the right side).
  - On Tablet, only the menu and the central content.
  - On Mobile, only the central content.

# Tests

I just placed an example for the menu to visually see.


https://github.com/user-attachments/assets/b80025bf-f3d9-4383-93ae-8cc9d92a3fcc

Slowed transition:



https://github.com/user-attachments/assets/3aeb6670-6556-4c5c-b2c5-62679c143d37

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>