style(frontend): adjust layout to have sticky header footer and menu [GIX-3046] (#2876)
# Motivation

We aim to leave the Header, the Footer and the Navigational Menu sticky
for medium and larger screens. However, the main content should be able
to freely scroll in the space occupied by the Header and the Footer,
when it not overlaps (basically larger screens). Meaning that the class
`sticky` is not apt to be used in that case.

So, we use `fixed` for all three, but being careful that it should
applied only when signed in (the landing page is ok as it is).

# Results


https://github.com/user-attachments/assets/88c2375c-6d27-4a91-8725-70aed20181ab

# Landing page unchanged



https://github.com/user-attachments/assets/a098e72d-c7c0-43fe-9572-298dd5734afa