feat(frontend): handle BTC latest block height endpoint error (#2896)
# Motivation

Similar as with fetching BTC txs data, we don't want worker to fail if
the latest block height cannot be fetched. If this value is not
available, it also means that we cannot properly parse transactions
response. For now, I added a simple check, but in the future we might
want to handle this case more gracefully (e.g. we still display txs but
mention to users that we cannot identify confirmations number and hence
tx status).