feat(frontend): signer context (#2934)
# Motivation

To implement the OISY Wallet Signer features, a consumer must create a
`Signer` object and implement various prompts.

In the OISY Wallet integration, I ended up using a context to handle all
subscriptions instead of, for example, registering prompts at the
component level. This approach felt cleaner and more aligned with our
general architecture.

I also did not declare a global store, but rather a context. This way,
we can potentially reuse the signer features more easily.

# Notes

This PR itself might not make much sense, but it is part of the branch
split from #2293, which might provide more context.

# Changes

- Create `signer.store` context initializer and interfaces.

# Tests

The tests are minimal. I tried to implement more features, such as
ensuring that the various prompts are triggered upon receiving Signer
Standards window post messages, but ultimately, it felt like an
overcomplication and a bit of a reinventing the wheel, particularly
since the library itself is already well-covered by tests. What we want
to assert here is that the components (in upcoming PRs) behave as they
should, and this can be validated through E2E tests.