feat(frontend): signer idle (#2950)
# Motivation

When the signer is idle - i.e. awaiting messages and not processing
anything - we want to display an idle related animation.

# Notes

This can notably happens if relying parties do not close the OISY Wallet
window after interactions.

I was hesitant to place one the new animated component in icon folder
but, ultimately felt like it was something related to the signer only
for now. So provided it as a `Signer...` component as well.

# Changes

- New idle component and animation

# Screenshot


![demo](https://github.com/user-attachments/assets/92a68844-5b31-4b22-ab54-02bd3abd4997)