feat(frontend): signer permissions (#2953)
# Motivation

The OISY Wallet Signer library handles requests and permission needs. On
the OISY Wallet side, we need to manage prompt requests, present the
information to the user, and provide them with a way to reject or
approve the permissions.

# Notes

While the integration with the demo begins by requesting all
permissions, according to the specification, individual permissions can
be requested separately. That’s why the component can also be triggered
independently.

The wallet address is section is rendered only if the related "accounts"
permissions are requested.

# Changes

- New component `SignerPermissions` which uses the prompt and provide
the call to action for the user
- Integrate the component in the `/sign` route

# Screenshot

<img width="688" alt="Capture d’écran 2024-10-16 à 18 35 28"
src="https://github.com/user-attachments/assets/a76ffde0-ae99-49f3-97eb-8da229cbf115">