feat(frontend): signer call canister info (#2960)
# Motivation

All the heavy work is handled by the OISY Wallet Signer library, so the
"Call Canister" component is mainly responsible for managing the related
prompts to display the status of the execution to the user.

# Notes

I'm not satisfied with the name "loading" for the state. We might rename
it soon. Essentially, "loading" means the call is being executed.

The recommended UX for the relying party is to close the modal, but they
might batch operations. We'll need to gather feedback to improve the UI.
In the meantime, we display a simple success or error result page.

# Changes

- Create a component to display the status of the call canister and it's
result
- Integrate component in `/sign` route
- Get rid of the "work in progress" text. This is the last PR to
integrate the Signer feature in OISY! 🥳

# Screenshot

<img width="688" alt="Capture d’écran 2024-10-16 à 18 52 42"
src="https://github.com/user-attachments/assets/9552372e-f8ac-4c33-aaa3-1bb5100ed852">
<img width="688" alt="Capture d’écran 2024-10-16 à 18 52 25"
src="https://github.com/user-attachments/assets/44e22be1-8ef0-44a8-b0d9-ceadc4afe89d">