fix(frontend): Clickable header and footer when signed in (#2959)
# Motivation

When signed in and in large screens, a part of the token list is
overlapped by the Header and the Footer. That means that those part are
un-clickable (for example Alpha and Manage list).

To solve it, we make the entire Header and Footer with disabled pointer
events, enabling it ONLY for the elements inside it.