feat(frontend): Use page token for balance in Hero (#2979)
# Motivation

In `HeroContent`, the component `Balance` was using the `balance` store,
that it is a derived of the `token` store by itself. However, in the
context of the usage of `Balance` component, we don't need to rely on
the `token` store, and we can use directly the `pageToken` store.

In fact, in this context, the token store is set as such:
token.set($pageToken).

So, we adapt `pageToken` store to have the balance values too.

# Changes

- Create new variable `pageTokenUi` in HeroContent that maps `pageToken`
to have financial data.
- Make `Balance` component accepts a `token` prop.
- Adapt Balance component to use the token prop instead of global
stores/derived.

# Tests

No changes in the local replica, same values showed.