refactor(frontend): extract copy to clipboard to a module (#3014)
# Motivation

In #3013 we will need a specific copy button icon. I'm not sure if it
will be use elsewhere, that's why instead of enhancing the existing
`Copy` button, I rather like to create a specific "Receive related
button". Also why it's handy to extract the "copy to clipboard" function
to a module for reusability purposes.

# Changes

- Move existing copy to clipboard function to a utils.

# Tests

No particular tests.