feat(frontend): Introduce OisySymbol prop for appearence interface (#3053)
# Motivation

For grouping tokens, we are going to re-use the `TokenCard` component
that takes a few props of a given token, for example `name` and
`symbol`. However, for the group, we will provide some different data to
be shown, just for appearance sake.

For prop `name`, we already have `oisyName` substitute to show something
different in place. But, for `symbol`, there is nothing.

So, similar to `oisyName`, we create an optional prop `oisySymbol` to be
provided if we want to show some substitute to the usual `token.symbol`,
just for different appearance.

UNRELATED: Since we are here, we will set `oisyName.prefix` as optional
prop, since it makes sense