feat(frontend): Include balances in token group (#3062)
# Motivation

Another useful PR for grouping tokens: we need to have the balances
(both token and USD) for each group. Note that for nullish values for
both tokens in group, we keep them as they are but prioritizing the
native token. So for example:

### Case 1

- native token --> `null`
- twin token --> `undefined`
- group --> `null`

### Case 2

- native token --> `undefined`
- twin token --> `null`
- group --> `undefined`

# Changes

- Change `TokenGroupUi` to be a type instead of interface and add the
financial data props.
- Add balances calculation to `createTokenGroup`.

# Tests

No changes in local replica, since we are not using those values yet.