feat(frontend): Carousel utils (#3067)
# Motivation

This PR includes utils needed for the Carousel component. The idea is to
take provided to `Carousel` as children slides (can be any HTML
elements, components, etc.) and to create a slider frame - a div that
has `slidesNumber * slideWidth` width (only active slide is visible) and
includes all slides with equal (percentage) size. The other 3 functions
are small helpers to enable/disable (we need to disable animation when
we do last-to-first and first-to-last switches) animation and to change
a current slide by applying `transform` to the frame.

Please let me know if you need more context to review this PR (e.g. you
also want to include the UI component itself here).



https://github.com/user-attachments/assets/0cf5618b-1e26-45d2-8b9f-b2d1c3ce771c