feat(frontend): Card component for Token Group (#3054)
# Motivation

We want to create a custom Token Card but for groups of tokens.
Basically, this card will collect a group of tokens and once clicked, it
will expand to show them. While it is collapsed, it will have the same
aspects of the other Token Cards.

# Note

Before being used, we MUST create a listener for all the tokens in the
group. We cannot create it inside the `IF` condition, otherwise the
summary will not be updated (since the `Listener` will not be mounted)

# Changes

- Create type for LogoSize, so that we can pass it as parameter and
bubble it up.
- Add prop `logoSize` in components `TokenLogo`, `TokenCard` and
`TokenCardContent`.
- Change `TokenCard` to accept a number of tokens: this prop will be
used if the `TokenCard` is acting as a summary card for the group.
- Create component TokenGroupCard that has:
  - A main "summary" card that re-uses the normal `TokenCard`
- Some sub-`TokenCard` that are built/shown when the main one is clicked
- We pass a custom created "summary" token to the "summary" token card,
that has all the properties of the Native Token of the group plus a few
changed properties for appearance and the financial data summirized.



# Tests

I mocked a token group, uniting ICP with BTC and giving fake data.


https://github.com/user-attachments/assets/0235bef3-cd11-44f0-97c2-162e71f57191



https://github.com/user-attachments/assets/c8e4ce34-4b2b-41ea-a8ef-da5c1f96a60b