feat(frontend): Create listener for multiple tokens (#3040)
# Motivation

The `TokenGroupCard` component created in PR #3054 is still lacking the
listeners for the tokens.

Basically, some tokens requires the `Listener` component to update the
balance. We cannot put it in the single-token cards that are
sub-elements of the group card, because they will not be mounted if the
group is not expanded.

So, to have the correct `Listener` objects, we will not wrap the cards
with the `Listener`, but we wrap the entire Token Group card with a new
component `MultipleListeners`.

This component is simply a stack of single-token `Listeners`, but only
the ones that are required.

# Changes

- Create util to map a Listener to a specific token: `BTC` -->
`BtcListener`, for example.
- Create component MultipleListener to stack the required `Listeners`
given a list of tokens.
- Wrap the token group card with the new `Listener` component.

# Test

Created some tests for the util function.