feat(frontend): Implement token groups in tokens list (#3099)
# Motivation

We want to start grouping tokens that refers to the same underlying. For
example `BTC` and `ckBTC`, or `ETH` and `ckETH`. Basically twin tokens,
where the "native token" definition falls on the most mainstream one of
the group (BTC, ETH, etc)

To do that we first group the sorted tokens and then we show
single-tokens differently from the token groups (we use the
`TokenCardGroup` component introduced recently).


# Changes

- Modify function `isTokenUiGroup` to recognize the correct group type
instead of the token type.
- Add an ID for the token group: this will be the same of the native
token so that the animation can recognize it as the same element.
- Group tokens with function `groupTokensByTwin` in
`TokenDisplayHandler`.
- In the token list, differentiate how we show a group from how we show
a single-token card.

# Tests

I deployed the current branch in a test environment (otherwise there
were no tokens to be grouped...):



https://github.com/user-attachments/assets/46f4abbc-e661-4e1b-aadf-18e973da1d97

---------

Co-authored-by: Dario Lötscher <dario.loetscher@inacta.ch>
Co-authored-by: darioloetsc <76442731+darioloetsc@users.noreply.github.com>