feat(frontend): adjust Carousel initialisation function (#3111)
# Motivation

While implementing Carousel usages (2 instances rendered on the token
page to comply with the provided design, one for XL screens, another one
for smaller ones), I noticed that the frame and timers are being
initialised even if one of carousel is not actually visible (display:
none). In this PR, I add a check to avoid that. The check works and
stops all processes both on mount and on resize.